@@ -1,6 +1,6 @@
 {
   "name": "@storybook/addon-actions",
-  "version": "3.3.0-alpha.0",
+  "version": "3.3.0-alpha.2",
   "description": "Action Logger addon for storybook",
   "keywords": [
     "storybook"
@@ -21,17 +21,16 @@
     "storybook": "start-storybook -p 9001"
   },
   "dependencies": {
-    "@storybook/addons": "^3.3.0-alpha.0",
+    "@storybook/addons": "^3.3.0-alpha.2",
     "deep-equal": "^1.0.1",
-    "json-stringify-safe": "^5.0.1",
-    "prop-types": "^15.5.10",
-    "react-inspector": "^2.1.1",
+    "prop-types": "^15.6.0",
+    "react-inspector": "^2.2.0",
     "uuid": "^3.1.0"
   },
   "devDependencies": {
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
-    "react-test-renderer": "^15.6.1",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
+    "react-test-renderer": "^16.0.0",
     "shelljs": "^0.7.8"
   },
   "peerDependencies": {
