@@ -1,6 +1,6 @@
 {
   "name": "@storybook/addon-graphql",
-  "version": "3.3.0-alpha.0",
+  "version": "3.3.0-alpha.2",
   "description": "Storybook addon to display the GraphiQL IDE",
   "keywords": [
     "storybook"
@@ -22,13 +22,13 @@
   },
   "dependencies": {
     "global": "^4.3.2",
-    "graphiql": "^0.7.8",
-    "graphql": "^0.7.0",
-    "prop-types": "^15.5.10"
+    "graphiql": "^0.11.6",
+    "graphql": "^0.11.7",
+    "prop-types": "^15.6.0"
   },
   "devDependencies": {
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
     "shelljs": "^0.7.8"
   },
   "peerDependencies": {
