@@ -1,6 +1,6 @@
 {
   "name": "@storybook/addon-knobs",
-  "version": "3.3.0-alpha.0",
+  "version": "3.3.0-alpha.2",
   "description": "Storybook Addon Prop Editor Component",
   "license": "MIT",
   "main": "dist/index.js",
@@ -11,34 +11,30 @@
   "scripts": {
     "prepare": "node ../../scripts/prepare.js",
     "publish-storybook": "bash .scripts/publish_storybook.sh",
-    "start": "./example/prepare.sh",
     "storybook": "start-storybook -p 9010"
   },
   "dependencies": {
-    "@storybook/addons": "^3.3.0-alpha.0",
-    "babel-runtime": "^6.23.0",
+    "@angular/core": "^5.0.0-beta.7",
+    "@storybook/addons": "^3.3.0-alpha.2",
+    "babel-runtime": "^6.26.0",
     "deep-equal": "^1.0.1",
     "global": "^4.3.2",
     "insert-css": "^1.0.0",
     "lodash.debounce": "^4.0.8",
-    "moment": "^2.18.1",
-    "prop-types": "^15.5.10",
+    "moment": "^2.19.1",
+    "prop-types": "^15.6.0",
     "react-color": "^2.11.4",
-    "react-datetime": "^2.8.10",
+    "react-datetime": "^2.10.3",
     "react-textarea-autosize": "^4.3.0",
     "util-deprecate": "^1.0.2"
   },
   "devDependencies": {
-    "@types/node": "^7.0.12",
-    "@types/react": "^15.0.21",
     "git-url-parse": "^6.2.2",
     "raw-loader": "^0.5.1",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
-    "style-loader": "^0.17.0",
-    "typescript": "^2.2.2",
-    "typescript-definition-tester": "^0.0.5",
-    "vue": "^2.4.2"
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
+    "style-loader": "^0.19.0",
+    "vue": "^2.5.2"
   },
   "peerDependencies": {
     "react": "*",
