@@ -1,6 +1,6 @@
 {
   "name": "@storybook/addon-notes",
-  "version": "3.3.0-alpha.0",
+  "version": "3.3.0-alpha.2",
   "description": "Write notes for your Storybook stories.",
   "keywords": [
     "addon",
@@ -19,16 +19,16 @@
     "storybook": "start-storybook -p 9010"
   },
   "dependencies": {
-    "@storybook/addons": "^3.3.0-alpha.0",
-    "babel-runtime": "^6.23.0",
-    "prop-types": "^15.5.10",
+    "@storybook/addons": "^3.3.0-alpha.2",
+    "babel-runtime": "^6.26.0",
+    "prop-types": "^15.6.0",
     "util-deprecate": "^1.0.2"
   },
   "devDependencies": {
     "git-url-parse": "^6.2.2",
-    "react": "^15.6.1",
+    "react": "^16.0.0",
     "react-addons-test-utils": "^15.5.1",
-    "react-dom": "^15.6.1"
+    "react-dom": "^16.0.0"
   },
   "peerDependencies": {
     "react": "*"
