@@ -1,6 +1,6 @@
 {
   "name": "@storybook/addon-options",
-  "version": "3.3.0-alpha.0",
+  "version": "3.3.0-alpha.2",
   "description": "Options addon for storybook",
   "keywords": [
     "storybook"
@@ -20,12 +20,12 @@
     "storybook": "start-storybook -p 9001"
   },
   "dependencies": {
-    "@storybook/addons": "^3.3.0-alpha.0"
+    "@storybook/addons": "^3.3.0-alpha.2"
   },
   "devDependencies": {
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
-    "react-test-renderer": "^15.6.1",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
+    "react-test-renderer": "^16.0.0",
     "shelljs": "^0.7.8"
   },
   "peerDependencies": {
