@@ -1,6 +1,6 @@
 {
   "name": "@storybook/addon-storyshots",
-  "version": "3.3.0-alpha.0",
+  "version": "3.3.0-alpha.2",
   "description": "StoryShots is a Jest Snapshot Testing Addon for Storybook.",
   "license": "MIT",
   "main": "dist/index.js",
@@ -15,17 +15,17 @@
     "example": "jest storyshot.test"
   },
   "dependencies": {
-    "babel-runtime": "^6.23.0",
+    "babel-runtime": "^6.26.0",
     "glob": "^7.1.2",
     "global": "^4.3.2",
     "jest-specific-snapshot": "^0.2.0",
-    "prop-types": "^15.5.10",
+    "prop-types": "^15.6.0",
     "read-pkg-up": "^2.0.0"
   },
   "devDependencies": {
-    "@storybook/addons": "^3.3.0-alpha.0",
-    "@storybook/channels": "^3.3.0-alpha.0",
-    "@storybook/react": "^3.3.0-alpha.0",
+    "@storybook/addons": "^3.3.0-alpha.2",
+    "@storybook/channels": "^3.3.0-alpha.2",
+    "@storybook/react": "^3.3.0-alpha.2",
     "babel-cli": "^6.26.0",
     "babel-jest": "^20.0.3",
     "babel-plugin-transform-runtime": "^6.23.0",
@@ -33,13 +33,13 @@
     "babel-preset-react": "^6.24.1",
     "jest": "^20.0.4",
     "jest-cli": "^20.0.4",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1"
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0"
   },
   "peerDependencies": {
-    "@storybook/addons": "^3.3.0-alpha.0",
-    "@storybook/channels": "^3.3.0-alpha.0",
-    "@storybook/react": "^3.3.0-alpha.0",
+    "@storybook/addons": "^3.3.0-alpha.2",
+    "@storybook/channels": "^3.3.0-alpha.2",
+    "@storybook/react": "^3.3.0-alpha.2",
     "babel-core": "^6.26.0",
     "react": "*",
     "react-test-renderer": "*"
