@@ -1,4 +1,4 @@
-const spawnSync = require('cross-spawn').sync;
+import { sync as spawnSync } from 'cross-spawn';
 
 export default function hasYarn() {
   const result = spawnSync('yarn', ['--version'], { silent: true });
