@@ -1,10 +1,41 @@
 import path from 'path';
 import fs from 'fs';
 import chalk from 'chalk';
+import latestVersion from 'latest-version';
+import { sync as spawnSync } from 'cross-spawn';
+import { gt } from 'semver';
 
-const spawnSync = require('cross-spawn').sync;
+import { version, devDependencies } from '../package.json';
 
 const logger = console;
+
+export async function getVersion(packageName) {
+  let current;
+  if (packageName === '@storybook/cli') {
+    current = version;
+  } else if (/storybook/.test(packageName)) {
+    current = devDependencies[packageName];
+  }
+
+  let latest;
+  try {
+    latest = await latestVersion(packageName);
+  } catch (e) {
+    if (current) {
+      logger.warn(`\n     ${chalk.yellow(e.message)}`);
+      return current;
+    }
+
+    logger.error(`\n     ${chalk.red(e.message)}`);
+    process.exit(1);
+  }
+
+  return current && gt(current.replace(/^\^/, ''), latest) ? current : `^${latest}`;
+}
+
+export function getVersions(...packageNames) {
+  return Promise.all(packageNames.map(getVersion));
+}
 
 export function getPackageJson() {
   const packageJsonPath = path.resolve('package.json');
@@ -21,6 +52,23 @@ export function writePackageJson(packageJson) {
   const packageJsonPath = path.resolve('package.json');
 
   fs.writeFileSync(packageJsonPath, content, 'utf8');
+}
+
+export function getBabelRc() {
+  const babelRcPath = path.resolve('.babelrc');
+  if (!fs.existsSync(babelRcPath)) {
+    return false;
+  }
+
+  const babelRcContent = fs.readFileSync(babelRcPath, 'utf8');
+  return JSON.parse(babelRcContent);
+}
+
+export function writeBabelRc(babelRc) {
+  const content = `${JSON.stringify(babelRc, null, 2)}\n`;
+  const babelRcPath = path.resolve('.babelrc');
+
+  fs.writeFileSync(babelRcPath, content, 'utf8');
 }
 
 export function commandLog(message) {
