@@ -15,6 +15,9 @@ module.exports = {
     node: true,
     'jest/globals': true,
   },
+  settings: {
+    'import/core-modules': ['enzyme'],
+  },
   rules: {
     strict: [error, 'never'],
     'prettier/prettier': [
@@ -31,10 +34,9 @@ module.exports = {
     'class-methods-use-this': ignore,
     'arrow-parens': [warn, 'as-needed'],
     'space-before-function-paren': ignore,
-    'import/no-unresolved': warn,
+    'import/no-unresolved': error,
     'import/extensions': [
-      // because of highlight.js and fuse.js
-      warn,
+      error,
       {
         js: 'never',
         json: 'always',
