@@ -15,7 +15,8 @@ class ActionLogger extends Component {
         <div style={style.countwrap}>{action.count > 1 && counter}</div>
         <div style={style.inspector}>
           <Inspector
-            showNonenumerable
+            sortObjectKeys
+            showNonenumerable={false}
             name={action.data.name}
             data={action.data.args || action.data}
           />
