@@ -46,16 +46,16 @@ export default class PropField extends React.Component {
   }
 
   render() {
-    const { onChange, knob } = this.props;
+    const { onChange, onClick, knob } = this.props;
 
     const InputType = TypeMap[knob.type] || InvalidType;
 
     return (
       <div style={stylesheet.field}>
         <label htmlFor={knob.name} style={stylesheet.label}>
-          {`${knob.name}`}
+          {!knob.hideLabel && `${knob.name}`}
         </label>
-        <InputType knob={knob} onChange={onChange} />
+        <InputType knob={knob} onChange={onChange} onClick={onClick} />
       </div>
     );
   }
@@ -67,4 +67,5 @@ PropField.propTypes = {
     value: PropTypes.any,
   }).isRequired,
   onChange: PropTypes.func.isRequired,
+  onClick: PropTypes.func.isRequired,
 };
