@@ -6,7 +6,7 @@ import KnobStore from '../KnobStore';
 describe('React Handler', () => {
   describe('wrapStory', () => {
     it('should contain the story and add correct props', () => {
-      const testChannel = { emit: () => {} };
+      const testChannel = { emit: jest.fn(), on: jest.fn() };
       const testStory = () => <div id="test-story">Test Content</div>;
       const testContext = {
         kind: 'Foo',
