@@ -4,7 +4,7 @@ import KnobStore from '../KnobStore';
 
 describe('Vue handler', () => {
   it('Returns a component with a created function', () => {
-    const testChannel = { emit: () => {} };
+    const testChannel = { emit: jest.fn(), on: jest.fn() };
     const testStory = () => ({ template: '<div> testStory </div>' });
     const testContext = {
       kind: 'Foo',
@@ -32,8 +32,9 @@ describe('Vue handler', () => {
     const testStore = new KnobStore();
     new Vue(vueHandler(testChannel, testStore)(testStory)(testContext)).$mount();
 
-    expect(testChannel.on).toHaveBeenCalledTimes(2);
+    expect(testChannel.on).toHaveBeenCalledTimes(3);
     expect(testChannel.on).toHaveBeenCalledWith('addon:knobs:reset', expect.any(Function));
     expect(testChannel.on).toHaveBeenCalledWith('addon:knobs:knobChange', expect.any(Function));
+    expect(testChannel.on).toHaveBeenCalledWith('addon:knobs:knobClick', expect.any(Function));
   });
 });
