@@ -1,6 +1,6 @@
 import path from 'path';
 
-export function getStoryshotFile(fileName) {
+function getStoryshotFile(fileName) {
   const { dir, name } = path.parse(fileName);
   return path.format({ dir: path.join(dir, '__snapshots__'), name, ext: '.storyshot' });
 }
@@ -12,4 +12,14 @@ export function getPossibleStoriesFiles(storyshotFile) {
     path.format({ dir: path.dirname(dir), name, ext: '.js' }),
     path.format({ dir: path.dirname(dir), name, ext: '.jsx' }),
   ];
+}
+
+export function getSnapshotFileName(context) {
+  const fileName = context.fileName;
+
+  if (!fileName) {
+    return null;
+  }
+
+  return getStoryshotFile(fileName);
 }
