@@ -1,4 +1,5 @@
 #!/usr/bin/env node
+/* eslint-disable no-console */
 
 import path from 'path';
 import program from 'commander';
@@ -16,6 +17,9 @@ program
   .option('--skip-packager', 'run only storybook server')
   .option('-i, --manual-id', 'allow multiple users to work with same storybook')
   .option('--smoke-test', 'Exit after successful start')
+  .option('--packager-port <packagerPort>', 'Custom packager port')
+  .option('--root [root]', 'Add additional root(s) to be used by the packager in this project')
+  .option('--projectRoots [projectRoots]', 'Override the root(s) to be used by the packager')
   .parse(process.argv);
 
 const projectDir = path.resolve();
@@ -38,7 +42,7 @@ server.listen(...listenAddr, err => {
     throw err;
   }
   const address = `http://${program.host || 'localhost'}:${program.port}/`;
-  console.info(`\nReact Native Storybook started on => ${address}\n`); // eslint-disable-line no-console
+  console.info(`\nReact Native Storybook started on => ${address}\n`);
   if (program.smokeTest) {
     process.exit(0);
   }
@@ -47,6 +51,12 @@ server.listen(...listenAddr, err => {
 if (!program.skipPackager) {
   let symlinks = [];
 
+  let roots = [projectDir];
+
+  if (program.root) {
+    roots = roots.concat(program.root.split(',').map(root => path.resolve(root)));
+  }
+
   try {
     const findSymlinksPaths = require('react-native/local-cli/util/findSymlinksPaths'); // eslint-disable-line global-require
     symlinks = findSymlinksPaths(path.join(projectDir, 'node_modules'), [projectDir]);
@@ -54,9 +64,15 @@ if (!program.skipPackager) {
     console.warn(`Unable to load findSymlinksPaths: ${e.message}`);
   }
 
-  const projectRoots = (configDir === projectDir ? [configDir] : [configDir, projectDir]).concat(
+  let projectRoots = (configDir === projectDir ? [configDir] : [configDir, projectDir]).concat(
     symlinks
   );
+
+  if (program.projectRoots) {
+    projectRoots = projectRoots.concat(
+      program.projectRoots.split(',').map(root => path.resolve(root))
+    );
+  }
 
   let cliCommand = 'node node_modules/react-native/local-cli/cli.js start';
   if (program.haul) {
@@ -67,8 +83,9 @@ if (!program.skipPackager) {
     [
       cliCommand,
       `--projectRoots ${projectRoots.join(',')}`,
-      `--root ${projectDir}`,
+      `--root ${roots.join(',')}`,
       program.resetCache && '--reset-cache',
+      program.packagerPort && `--port=${program.packagerPort}`,
     ]
       .filter(x => x)
       .join(' '),
