@@ -2,7 +2,5 @@ import deprecate from 'util-deprecate';
 import '@storybook/addon-actions/register';
 import '@storybook/addon-links/register';
 
-deprecate(
-  () => {},
-  '@storybook/react/addons is deprecated. See https://storybook.js.org/addons/using-addons/'
-)();
+deprecate(() => {},
+'@storybook/react/addons is deprecated. See https://storybook.js.org/addons/using-addons/')();
