@@ -1,1 +1,3 @@
-require('airbnb-js-shims');
+import 'core-js/es6/symbol';
+import 'core-js/fn/array/iterator';
+import 'airbnb-js-shims';
