@@ -150,10 +150,16 @@ server.listen(...listenAddr, error => {
 
 Promise.all([webpackValid, serverListening])
   .then(() => {
-    const address = `http://${program.host || 'localhost'}:${program.port}/`;
+    const proto = program.https ? 'https' : 'http';
+    const address = `${proto}://${program.host || 'localhost'}:${program.port}/`;
     logger.info(`Storybook started on => ${chalk.cyan(address)}\n`);
     if (program.smokeTest) {
       process.exit(0);
     }
   })
-  .catch(error => logger.error(error));
+  .catch(error => {
+    logger.error(error);
+    if (program.smokeTest) {
+      process.exit(1);
+    }
+  });
