@@ -1,11 +1,12 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import EventEmiter from 'eventemitter3';
 
 import { storiesOf } from '@storybook/react';
 import { setOptions } from '@storybook/addon-options';
 import { action } from '@storybook/addon-actions';
 import { withNotes, WithNotes } from '@storybook/addon-notes';
-import { linkTo } from '@storybook/addon-links';
+import { LinkTo, linkTo, hrefTo } from '@storybook/addon-links';
 import WithEvents from '@storybook/addon-events';
 import {
   withKnobs,
@@ -16,6 +17,7 @@ import {
   select,
   array,
   date,
+  button,
   object,
 } from '@storybook/addon-knobs/react';
 import centered from '@storybook/addon-centered';
@@ -40,7 +42,7 @@ const EVENTS = {
 const emiter = new EventEmiter();
 const emit = emiter.emit.bind(emiter);
 
-storiesOf('Welcome', module).add('to Storybook', () => <Welcome showApp={linkTo('Button')} />);
+storiesOf('Welcome', module).add('to Storybook', () => <Welcome showKind="Button" />);
 
 const InfoButton = () => (
   <span
@@ -59,6 +61,23 @@ const InfoButton = () => (
     Show Info{' '}
   </span>
 );
+
+class AsyncItemLoader extends React.Component {
+  constructor() {
+    super();
+    this.state = { items: [] };
+  }
+
+  loadItems() {
+    setTimeout(() => this.setState({ items: ['pencil', 'pen', 'eraser'] }), 1500);
+  }
+
+  render() {
+    button('Load the items', () => this.loadItems());
+    return this.props.children(this.state.items);
+  }
+}
+AsyncItemLoader.propTypes = { children: PropTypes.func.isRequired };
 
 storiesOf('Button', module)
   .addDecorator(withKnobs)
@@ -137,6 +156,14 @@ storiesOf('Button', module)
         <p>In my backpack, I have:</p>
         <ul>{items.map(item => <li key={item}>{item}</li>)}</ul>
         <p>{salutation}</p>
+        <hr />
+        <p>PS. My shirt pocket contains: </p>
+        <AsyncItemLoader>
+          {loadedItems => {
+            if (!loadedItems.length) return <li>No items!</li>;
+            return <ul>{loadedItems.map(i => <li key={i}>{i}</li>)}</ul>;
+          }}
+        </AsyncItemLoader>
       </div>
     );
   })
@@ -172,25 +199,52 @@ storiesOf('Button', module)
     )
   );
 
+storiesOf('AddonLink.Link', module)
+  .add('First', () => <LinkTo story="Second">Go to Second</LinkTo>)
+  .add('Second', () => <LinkTo story="First">Go to First</LinkTo>);
+
+storiesOf('AddonLink.Button', module)
+  .add('First', () => (
+    <button onClick={linkTo('AddonLink.Button', 'Second')}>Go to "Second"</button>
+  ))
+  .add('Second', () => (
+    <button onClick={linkTo('AddonLink.Button', 'First')}>Go to "First"</button>
+  ));
+
+storiesOf('AddonLink.Select', module)
+  .add('Index', () => (
+    <select value="Index" onChange={linkTo('AddonLink.Select', e => e.currentTarget.value)}>
+      <option>Index</option>
+      <option>First</option>
+      <option>Second</option>
+      <option>Third</option>
+    </select>
+  ))
+  .add('First', () => <LinkTo story="Index">Go back</LinkTo>)
+  .add('Second', () => <LinkTo story="Index">Go back</LinkTo>)
+  .add('Third', () => <LinkTo story="Index">Go back</LinkTo>);
+
+storiesOf('AddonLink.Href', module).add('log', () => {
+  hrefTo('AddonLink.Href', 'log').then(action('URL of this story'));
+
+  return <span>See action logger</span>;
+});
+
 storiesOf('AddonInfo.DocgenButton', module).addWithInfo('DocgenButton', 'Some Description', () => (
   <DocgenButton onClick={action('clicked')} label="Docgen Button" />
 ));
 
-storiesOf(
-  'AddonInfo.ImportedPropsButton',
-  module
-).addWithInfo(
+storiesOf('AddonInfo.ImportedPropsButton', module).addWithInfo(
   'ImportedPropsButton',
   'Button with PropTypes imported from another file. Should fallback to using PropTypes for data.',
   () => <ImportedPropsButton onClick={action('clicked')} label="Docgen Button" />
 );
 
-storiesOf(
-  'AddonInfo.FlowTypeButton',
-  module
-).addWithInfo('FlowTypeButton', 'Some Description', () => (
-  <FlowTypeButton onClick={action('clicked')} label="Flow Typed Button" />
-));
+storiesOf('AddonInfo.FlowTypeButton', module).addWithInfo(
+  'FlowTypeButton',
+  'Some Description',
+  () => <FlowTypeButton onClick={action('clicked')} label="Flow Typed Button" />
+);
 
 storiesOf('App', module).add('full app', () => <App />);
 
