@@ -22,5 +22,6 @@ module.exports = {
   ],
   coverageDirectory: 'coverage',
   testEnvironment: 'jsdom',
-  setupTestFrameworkScriptFile: './node_modules/jest-enzyme/lib/index.js',
+  setupTestFrameworkScriptFile: './scripts/jest.init.js',
+  setupFiles: ['raf/polyfill'],
 };
