@@ -1,16 +1,17 @@
 import mergeDirs from 'merge-dirs';
 import path from 'path';
 import shell from 'shelljs';
-import latestVersion from 'latest-version';
 import chalk from 'chalk';
-import { getPackageJson, writePackageJson, paddedLog } from '../../lib/helpers';
+import { getVersions, getPackageJson, writePackageJson, paddedLog } from '../../lib/helpers';
 
-export default Promise.all([
-  latestVersion('@storybook/react-native'),
-  latestVersion('@storybook/addon-actions'),
-  latestVersion('@storybook/addon-links'),
-  latestVersion('prop-types'),
-]).then(([storybookVersion, actionsVersion, linksVersion, propTypesVersion]) => {
+export default async () => {
+  const [storybookVersion, actionsVersion, linksVersion, propTypesVersion] = await getVersions(
+    '@storybook/react-native',
+    '@storybook/addon-actions',
+    '@storybook/addon-links',
+    'prop-types'
+  );
+
   // copy all files from the template directory to project directory
   mergeDirs(path.resolve(__dirname, 'template/'), '.', 'overwrite');
 
@@ -33,9 +34,9 @@ export default Promise.all([
   packageJson.dependencies = packageJson.dependencies || {};
   packageJson.devDependencies = packageJson.devDependencies || {};
 
-  packageJson.devDependencies['@storybook/react-native'] = `^${storybookVersion}`;
-  packageJson.devDependencies['@storybook/addon-actions'] = `^${actionsVersion}`;
-  packageJson.devDependencies['@storybook/addon-links'] = `^${linksVersion}`;
+  packageJson.devDependencies['@storybook/react-native'] = storybookVersion;
+  packageJson.devDependencies['@storybook/addon-actions'] = actionsVersion;
+  packageJson.devDependencies['@storybook/addon-links'] = linksVersion;
 
   if (!packageJson.dependencies['react-dom'] && !packageJson.devDependencies['react-dom']) {
     const reactVersion = packageJson.dependencies.react;
@@ -43,11 +44,11 @@ export default Promise.all([
   }
 
   if (!packageJson.dependencies['prop-types'] && !packageJson.devDependencies['prop-types']) {
-    packageJson.devDependencies['prop-types'] = `^${propTypesVersion}`;
+    packageJson.devDependencies['prop-types'] = propTypesVersion;
   }
 
   packageJson.scripts = packageJson.scripts || {};
   packageJson.scripts.storybook = 'storybook start -p 7007';
 
   writePackageJson(packageJson);
-});
+};
