@@ -1,23 +1,24 @@
 import mergeDirs from 'merge-dirs';
 import path from 'path';
 import fs from 'fs';
-import latestVersion from 'latest-version';
-import { getPackageJson, writePackageJson } from '../../lib/helpers';
+import { getVersions, getPackageJson, writePackageJson } from '../../lib/helpers';
 
-export default Promise.all([
-  latestVersion('@storybook/react'),
-  latestVersion('@storybook/addon-actions'),
-  latestVersion('@storybook/addon-links'),
-]).then(([storybookVersion, actionsVersion, linksVersion]) => {
+export default async () => {
+  const [storybookVersion, actionsVersion, linksVersion] = await getVersions(
+    '@storybook/react',
+    '@storybook/addon-actions',
+    '@storybook/addon-links'
+  );
+
   mergeDirs(path.resolve(__dirname, 'template/'), '.', 'overwrite');
 
   const packageJson = getPackageJson();
 
   packageJson.devDependencies = packageJson.devDependencies || {};
 
-  packageJson.devDependencies['@storybook/react'] = `^${storybookVersion}`;
-  packageJson.devDependencies['@storybook/addon-actions'] = `^${actionsVersion}`;
-  packageJson.devDependencies['@storybook/addon-links'] = `^${linksVersion}`;
+  packageJson.devDependencies['@storybook/react'] = storybookVersion;
+  packageJson.devDependencies['@storybook/addon-actions'] = actionsVersion;
+  packageJson.devDependencies['@storybook/addon-links'] = linksVersion;
 
   packageJson.scripts.storybook = 'start-storybook -p 9009';
   packageJson.scripts['build-storybook'] = 'build-storybook';
@@ -29,4 +30,4 @@ export default Promise.all([
   }
 
   writePackageJson(packageJson);
-});
+};
