@@ -7,9 +7,10 @@
 <script>
   export default {
     name: 'my-button',
-  
+
     methods: {
       onClick () {
+        this.$emit('click');
       }
     }
   }
