@@ -1,34 +1,32 @@
 import path from 'path';
-import latestVersion from 'latest-version';
 import { spawn } from 'child-process-promise';
 import { packageNames } from '@storybook/codemod';
-import { getPackageJson, writePackageJson } from '../../lib/helpers';
+import { getVersion, getPackageJson, writePackageJson } from '../../lib/helpers';
 
-function updatePackage(devDependencies, oldName, newName) {
+async function updatePackage(devDependencies, oldName, newName) {
   if (devDependencies[oldName]) {
-    return latestVersion(newName).then(version => {
-      delete devDependencies[oldName];  // eslint-disable-line
-      devDependencies[newName] = version; // eslint-disable-line
-    });
+    /* eslint-disable no-param-reassign */
+    delete devDependencies[oldName];
+    devDependencies[newName] = await getVersion(newName);
+    /* eslint-enable */
   }
-  return Promise.resolve(null);
 }
 
-function updatePackageJson() {
+async function updatePackageJson() {
   const packageJson = getPackageJson();
   const { devDependencies } = packageJson;
 
-  return Promise.all(
+  await Promise.all(
     Object.keys(packageNames).map(oldName => {
       const newName = packageNames[oldName];
       return updatePackage(devDependencies, oldName, newName);
     })
-  ).then(() => {
-    if (!devDependencies['@storybook/react'] && !devDependencies['@storybook/react-native']) {
-      throw new Error('Expected to find `@kadira/[react-native]-storybook` in devDependencies');
-    }
-    writePackageJson(packageJson);
-  });
+  );
+
+  if (!devDependencies['@storybook/react'] && !devDependencies['@storybook/react-native']) {
+    throw new Error('Expected to find `@kadira/[react-native]-storybook` in devDependencies');
+  }
+  writePackageJson(packageJson);
 }
 
 function updateSourceCode(parser) {
@@ -47,4 +45,7 @@ function updateSourceCode(parser) {
   return spawn(jscodeshiftCommand, args, { stdio: 'inherit' });
 }
 
-export default parser => updatePackageJson().then(() => updateSourceCode(parser));
+export default async parser => {
+  await updatePackageJson();
+  return updateSourceCode(parser);
+};
