@@ -1,4 +1,7 @@
+/* eslint-disable react/react-in-jsx-scope */
+
 import { storiesOf } from '@storybook/vue';
+import { action } from '@storybook/addon-actions';
 import { linkTo } from '@storybook/addon-links';
 
 import MyButton from './MyButton';
@@ -14,10 +17,19 @@ storiesOf('Button', module)
   .add('with text', () => ({
     components: { MyButton },
     template: '<my-button @click="action">Hello Button</my-button>',
+    methods: { action: action('clicked') },
+  }))
+  .add('with JSX', () => ({
+    components: { MyButton },
+    render() {
+      return <my-button onClick={this.action}>With JSX</my-button>;
+    },
     methods: { action: linkTo('clicked') },
   }))
   .add('with some emoji', () => ({
     components: { MyButton },
     template: '<my-button @click="action">😀 😎 👍 💯</my-button>',
-    methods: { action: linkTo('clicked') },
+    methods: { action: action('clicked') },
   }));
+
+/* eslint-enable react/react-in-jsx-scope */
