@@ -13,7 +13,8 @@ export default function detect(options) {
     if (
       packageJson.devDependencies['@storybook/react'] ||
       packageJson.devDependencies['@storybook/react-native'] ||
-      packageJson.devDependencies['@storybook/vue']
+      packageJson.devDependencies['@storybook/vue'] ||
+      packageJson.devDependencies['@storybook/angular']
     ) {
       return types.ALREADY_HAS_STORYBOOK;
     }
@@ -40,6 +41,7 @@ export default function detect(options) {
   if (
     (packageJson.dependencies && packageJson.dependencies.vue) ||
     (packageJson.devDependencies && packageJson.devDependencies.vue) ||
+    (packageJson.dependencies && packageJson.dependencies.nuxt) ||
     (packageJson.devDependencies && packageJson.devDependencies.nuxt)
   ) {
     return types.VUE;
@@ -81,6 +83,13 @@ export default function detect(options) {
     (packageJson.devDependencies && packageJson.devDependencies.react)
   ) {
     return types.REACT;
+  }
+
+  if (
+    (packageJson.dependencies && packageJson.dependencies['@angular/core']) ||
+    (packageJson.devDependencies && packageJson.devDependencies['@angular/core'])
+  ) {
+    return types.ANGULAR;
   }
 
   return types.UNDETECTED;
