@@ -32,11 +32,9 @@ export default class ReactProvider extends Provider {
                   inp = i;
                 }}
                 value={
-                  this.api.getQueryParam('text') === undefined ? (
-                    'ONE'
-                  ) : (
-                    this.api.getQueryParam('text')
-                  )
+                  this.api.getQueryParam('text') === undefined
+                    ? 'ONE'
+                    : this.api.getQueryParam('text')
                 }
                 onChange={() => {
                   this.api.setQueryParams({ text: inp.value });
