@@ -75,7 +75,8 @@ const fullScreenPreviewStyle = {
   border: 0,
   margin: 0,
   padding: 0,
-  overflow: 'hidden',
+  overflow: 'auto',
+  webkitOverflowScrolling: 'touch',
 };
 
 const overlayStyle = isDragging => ({
