@@ -6,7 +6,6 @@ const wrapperStyle = {
   background: '#F7F7F7',
   marginBottom: 10,
   display: 'flex',
-  height: 27,
 };
 
 const headingStyle = {
@@ -16,8 +15,9 @@ const headingStyle = {
   fontSize: '12px',
   fontWeight: 'bolder',
   color: '#828282',
+  textAlign: 'center',
   cursor: 'pointer',
-  padding: 0,
+  padding: '5px',
   margin: 0,
   overflow: 'hidden',
 };
@@ -37,6 +37,7 @@ const shortcutIconStyle = {
   backgroundColor: 'inherit',
   outline: 0,
   width: 30,
+  flexShrink: 0,
 };
 
 const linkStyle = {
