@@ -7,7 +7,7 @@ import { storyFilter } from '../../../libs/filters';
 
 const leftClick = { button: 0 };
 
-describe('manager.ui.components.stories_panel.stories', () => {
+describe('manager.ui.components.left_panel.stories', () => {
   beforeEach(() =>
     setContext({
       clientStore: {
@@ -48,6 +48,7 @@ describe('manager.ui.components.stories_panel.stories', () => {
           selectedKind={''}
           selectedStory={''}
           selectedHierarchy={[]}
+          sidebarAnimations={false}
         />
       );
 
@@ -67,6 +68,7 @@ describe('manager.ui.components.stories_panel.stories', () => {
           selectedKind="b"
           selectedStory="b2"
           selectedHierarchy={['b']}
+          sidebarAnimations={false}
         />
       );
 
@@ -83,6 +85,7 @@ describe('manager.ui.components.stories_panel.stories', () => {
           selectedKind="another.space.20"
           selectedStory="b2"
           selectedHierarchy={['another', 'space', '20']}
+          sidebarAnimations={false}
         />
       );
 
@@ -107,6 +110,7 @@ describe('manager.ui.components.stories_panel.stories', () => {
           selectedKind="another.space.20"
           selectedStory="b2"
           selectedHierarchy={['another.space.20']}
+          sidebarAnimations={false}
         />
       );
 
@@ -127,6 +131,7 @@ describe('manager.ui.components.stories_panel.stories', () => {
           selectedKind="another.space.20"
           selectedStory="b2"
           selectedHierarchy={['another', 'space', '20']}
+          sidebarAnimations={false}
         />
       );
 
@@ -146,10 +151,11 @@ describe('manager.ui.components.stories_panel.stories', () => {
           selectedKind="another.space.20"
           selectedStory="b2"
           selectedHierarchy={['another', 'space', '20']}
+          sidebarAnimations={false}
         />
       );
 
-      const kind = wrap.find('[data-name="some"]');
+      const kind = wrap.find('[data-name="some"]').first();
       kind.simulate('click', leftClick);
 
       const { nodes } = wrap.state();
@@ -169,6 +175,7 @@ describe('manager.ui.components.stories_panel.stories', () => {
           selectedKind="another.space.20"
           selectedStory="b2"
           selectedHierarchy={[]}
+          sidebarAnimations={false}
         />
       );
 
@@ -190,6 +197,7 @@ describe('manager.ui.components.stories_panel.stories', () => {
           selectedKind="another.space.20"
           selectedStory="b2"
           selectedHierarchy={['another', 'space', '20']}
+          sidebarAnimations={false}
         />
       );
 
@@ -215,6 +223,7 @@ describe('manager.ui.components.stories_panel.stories', () => {
           selectedKind="another.space.20"
           selectedStory="b2"
           selectedHierarchy={selectedHierarchy}
+          sidebarAnimations={false}
         />
       );
 
@@ -245,6 +254,7 @@ describe('manager.ui.components.stories_panel.stories', () => {
           selectedStory="b2"
           selectedHierarchy={['another', 'space', '20']}
           storyFilter={filter}
+          sidebarAnimations={false}
         />
       );
 
@@ -264,10 +274,11 @@ describe('manager.ui.components.stories_panel.stories', () => {
           selectedStory="b2"
           selectedHierarchy={['b']}
           onSelectStory={onSelectStory}
+          sidebarAnimations={false}
         />
       );
 
-      const kind = wrap.find('[data-name="a"]');
+      const kind = wrap.find('[data-name="a"]').first();
       kind.simulate('click', leftClick);
 
       expect(onSelectStory).not.toHaveBeenCalled();
@@ -282,6 +293,7 @@ describe('manager.ui.components.stories_panel.stories', () => {
           selectedStory="b2"
           selectedHierarchy={['b']}
           onSelectStory={onSelectStory}
+          sidebarAnimations={false}
         />
       );
 
@@ -306,10 +318,11 @@ describe('manager.ui.components.stories_panel.stories', () => {
           selectedStory="b2"
           selectedHierarchy={['b']}
           onSelectStory={onSelectStory}
+          sidebarAnimations={false}
         />
       );
 
-      const kind = wrap.find('[data-name="b1"]');
+      const kind = wrap.find('[data-name="b1"]').first();
       kind.simulate('click', leftClick);
 
       expect(onSelectStory).toHaveBeenCalledWith('b', 'b1');
@@ -324,16 +337,31 @@ describe('manager.ui.components.stories_panel.stories', () => {
           selectedStory="a2"
           selectedHierarchy={['some', 'name', 'item1']}
           onSelectStory={onSelectStory}
+          sidebarAnimations={false}
         />
       );
 
-      wrap.find('[data-name="another"]').simulate('click', leftClick);
-      wrap.find('[data-name="space"]').simulate('click', leftClick);
-      wrap.find('[data-name="20"]').simulate('click', leftClick);
+      wrap
+        .find('[data-name="another"]')
+        .first()
+        .simulate('click', leftClick);
+
+      wrap
+        .find('[data-name="space"]')
+        .first()
+        .simulate('click', leftClick);
+
+      wrap
+        .find('[data-name="20"]')
+        .first()
+        .simulate('click', leftClick);
 
       expect(onSelectStory).not.toHaveBeenCalled();
 
-      wrap.find('[data-name="b2"]').simulate('click', leftClick);
+      wrap
+        .find('[data-name="b2"]')
+        .first()
+        .simulate('click', leftClick);
 
       expect(onSelectStory).toHaveBeenCalledWith('another.space.20', 'b2');
     });
@@ -347,16 +375,32 @@ describe('manager.ui.components.stories_panel.stories', () => {
           selectedStory="a2"
           selectedHierarchy={['some', 'name', 'item1']}
           onSelectStory={onSelectStory}
+          sidebarAnimations={false}
         />
       );
 
-      wrap.find('[data-name="another"]').simulate('keyDown', { keyCode: 13 });
-      wrap.find('[data-name="space"]').simulate('keyDown', { keyCode: 13 });
-      wrap.find('[data-name="20"]').simulate('keyDown', { keyCode: 13 });
+      wrap
+        .find('[data-name="another"]')
+        .first()
+        .simulate('keyDown', { keyCode: 13 });
 
       expect(onSelectStory).not.toHaveBeenCalled();
 
-      wrap.find('[data-name="b2"]').simulate('click', leftClick);
+      wrap
+        .find('[data-name="space"]')
+        .first()
+        .simulate('keyDown', { keyCode: 13 });
+
+      // enter press on native link triggers click event
+      wrap
+        .find('[data-name="20"]')
+        .first()
+        .simulate('click', leftClick);
+
+      wrap
+        .find('[data-name="b2"]')
+        .first()
+        .simulate('click', leftClick);
 
       expect(onSelectStory).toHaveBeenCalledWith('another.space.20', 'b2');
     });
