@@ -23,9 +23,9 @@ describe('manager.ui.components.stories_panel.tree_decorators_utils.test', () =>
       const result = highlightNode(node, { highLightText: { color: 'red' } });
 
       expect(shallow(result[0]).html()).toEqual('<span>s</span>');
-      expect(shallow(result[1]).html()).toEqual('<strong style="color:red;">ome</strong>');
+      expect(shallow(result[1]).html()).toEqual('<strong style="color:red">ome</strong>');
       expect(shallow(result[2]).html()).toEqual('<span> </span>');
-      expect(shallow(result[3]).html()).toEqual('<strong style="color:red;">nam</strong>');
+      expect(shallow(result[3]).html()).toEqual('<strong style="color:red">nam</strong>');
       expect(shallow(result[4]).html()).toEqual('<span>e</span>');
     });
   });
