@@ -97,11 +97,11 @@ export default class TextFilter extends React.Component {
           />
         </div>
         {this.state.query &&
-        this.state.query.length && (
-          <button style={clearButtonStyle} onClick={this.fireOnClear} className="clear">
-            ×
-          </button>
-        )}
+          this.state.query.length && (
+            <button style={clearButtonStyle} onClick={this.fireOnClear} className="clear">
+              ×
+            </button>
+          )}
       </div>
     );
   }
