@@ -38,7 +38,7 @@ describe('manager.ui.components.stories_panel.test_filter', () => {
       const wrap = mount(<TextFilter onClear={onClear} />);
       wrap.setState({ query: 'hello' });
 
-      const clear = wrap.find('.clear');
+      const clear = wrap.find('.clear').first();
       clear.simulate('click');
 
       expect(onClear).toHaveBeenCalled();
